refactored command dialog
