+++ src/pages/main/cmdkPage/types.ts
@@ -1,0 +1,7 @@
+import { TCmdkReference } from "@_types/main";
+
+export type CommandItemProps = {
+  command: TCmdkReference;
+  index: number;
+  onSelect: (command: TCmdkReference) => Promise<void>;
+};
